+++ routes/manga.js
@@ -36,6 +36,9 @@ router.get("/search", [authenticate, rateLimti, logDB], async function(req,res){
         let resutls = [];
         const {q} = req.query;
         const result = await axios.get(`https://api.jikan.moe/v4/manga?q='${q}'`);
+        if(result.data.data.length <= 0){
+            return res.status(404).send({"message":"Manga tidak ditemukan!"})
+        }
         for (let i = 0;i<result.data.data.length;i++){
             const release_date = releaseDate(result.data.data[i].published.from);
             const r = {
@@ -73,7 +76,9 @@ router.get("/detail/:id", [authenticate, rateLimti, logDB], async function(req,r
         return res.status(200).send(item);
     }
     catch (err){
-        return res.status(500).send(err.toString());
+        return res.status(404).send({
+            "message":"Manga tidak ditemukan!"
+        });
     }
 });
 
@@ -83,6 +88,9 @@ router.get("/readList", [authenticate, rateLimti, logDB], async function(req,res
        const userdata = req.userdata;
        const mangaLists = [];
        const mangaList = await executeQuery(`select * from manga_lists where user_id = ${userdata.user_id}`);
+       if(mangaList.length <= 0){
+           return res.status(200).send({"message":"Tidak mempunyai list manga"})
+       }
        for(let i = 0;i<mangaList.length;i++){
            const a = {
                "list_id": mangaList[i].list_id,
@@ -145,6 +153,11 @@ router.put("/renameList", [authenticate, rateLimti, logDB], async function(req, 
        const {list_id, nama_list} = req.body;
        const mangaList = await executeQuery(`select * from manga_lists where list_id = ${list_id}`);
 
+       if(mangaList.length<=0){
+           return res.status(200).send({
+               "message":"List manga tidak ditemukan!"
+           })
+       }
        if(mangaList[0].user_id != userdata.user_id){
            return res.status(400).send({
                "message": "Bukan list anda!"
@@ -178,6 +191,16 @@ router.delete("/deleteList", [authenticate, rateLimti, logDB], async function(re
         const {list_id} = req.body;
 
         const mangaList = await executeQuery(`select * from manga_lists where list_id = ${list_id}`);
+        if(mangaList.length<=0){
+            return res.status(200).send({
+                "message":"List manga tidak ditemukan!"
+            })
+        }
+        if(mangaList[0].user_id != userdata.user_id){
+            return res.status(400).send({
+                "message": "Bukan list anda!"
+            });
+        }
         await executeQuery(`delete from manga_list_items where list_id = ${list_id}`);
         await executeQuery(`delete from manga_lists where list_id = ${list_id}`);
